requirements-local.txt
# Base
# CMAKE_ARGS="-DLLAMA_METAL=on" pip install -e <path/to/llama-cpp-python>
-e /Users/jethroestrada/Desktop/External_Projects/AI/repo-libs/llama-cpp-python
-e /Users/jethroestrada/Desktop/External_Projects/AI/repo-libs/litellm
-e /Users/jethroestrada/Desktop/External_Projects/AI/repo-libs/fastmcp
-e /Users/jethroestrada/Desktop/External_Projects/AI/repo-libs/deepeval
-e /Users/jethroestrada/Desktop/External_Projects/AI/repo-libs/BERTopic
-e /Users/jethroestrada/Desktop/External_Projects/AI/repo-libs/KeyBERT
-e /Users/jethroestrada/Desktop/External_Projects/AI/repo-libs/wtpsplit

# Advanced
-e /Users/jethroestrada/Desktop/External_Projects/AI/repo-libs/stanza
fake_useragent
validators
tiktoken
pyquery
parsel
html2text
justext
lxml
lxml-html-clean
Markdown
markdownify
mrkdwn-analysis
markdown-it-py
# for tables, tasklists, etc.
mdit-py-plugins
mdit-plain
model2vec
fastembed
wordcloud
# infinity_client

# KeyBERT Peers
# flair
